move links to another file
