+++ playground/plugins/trpc-client.ts
@@ -8,11 +8,11 @@ export default defineNuxtPlugin(() => {
     transformer: superjson,
     links: [
       // adds pretty logs to your console in development and logs errors in production
-      // loggerLink({
-      //   enabled: opts =>
-      //     process.env.NODE_ENV === 'development' ||
-      //     (opts.direction === 'down' && opts.result instanceof Error)
-      // }),
+      loggerLink({
+        enabled: opts =>
+          process.env.NODE_ENV === 'development' ||
+          (opts.direction === 'down' && opts.result instanceof Error)
+      }),
       httpBatchLink()
     ]
   })
